Refactored Process Manager, and splitted into Process Manager, Resource Manager and Plugin Manager
