+++ sqapi/core/process_manager.py
@@ -1,16 +1,12 @@
-import copy
 import logging
 import multiprocessing
-import os
 import threading
 
+from sqapi.core.plugin_manager import PluginManager
 from sqapi.core.message import Message
-from sqapi.core.processor import Processor
 from sqapi.query import data as q_data, metadata as q_meta
 from sqapi.util import detector
 from sqapi.util.cfg_util import Config
-
-SINGLE_PLUGIN = os.environ.get('PLUGIN', None)
 
 STATUS_VALIDATING = 'VALIDATING'
 STATUS_QUERYING = 'QUERYING'
@@ -23,57 +19,17 @@ log = logging.getLogger(__name__)
 
 
 class ProcessManager:
-    def __init__(self, config: Config):
+    def __init__(self, config: Config, plugin_manager: PluginManager):
         self.config = config
+        self.plugin_manager = plugin_manager
 
         self.database = detector.detect_database(self.config.database)
         self.database.initialize_message_table()
 
-        self.plugins = []
-        self.failed_plugins = []
-        self.register_plugins()
-
-        self.accepted_types = set()
-        for plugin in self.plugins:
-            supported_types = plugin.config.msg_broker.get('supported_mime') or ['*']
-            log.debug('Plugin {} accepts data types: {}'.format(plugin.name, supported_types))
-            log.debug('{} msg_broker config: {}'.format(plugin.name, plugin.config.msg_broker))
-            self.accepted_types.update(supported_types)
-
         self.listener = detector.detect_listener(self.config.msg_broker, self.process_message)
 
-    def register_plugins(self):
-        log.debug('Searching for available and active plugins')
-        detected_plugins = detector.detect_plugins().items()
-        for plugin_name, plugin in detected_plugins:
-            if not self.active_plugin(plugin_name):
-                log.debug('Plugin {} is not listed as active'.format(plugin_name))
-                continue
-
-            try:
-                log.debug('Registering a processor for plugin {}'.format(plugin_name))
-                processor = Processor(copy.deepcopy(self.config), plugin_name, plugin)
-                self.plugins.append(processor)
-
-            except Exception as e:
-                err = 'Could not register plugin {} ({}): {}'.format(plugin_name, plugin, str(e))
-                log.warning(err)
-                self.failed_plugins.append({
-                    plugin_name: err
-                })
-
-        log.info('{}/{} registered plugins'.format(len(self.plugins), len(detected_plugins)))
-        log.info('{}/{} failed plugins'.format(len(self.failed_plugins), len(detected_plugins)))
-        log.info('{}/{} inactive plugins'.format(len(self.plugins) + len(self.failed_plugins), len(detected_plugins)))
-
-    def active_plugin(self, plugin_name):
-        if SINGLE_PLUGIN:
-            return SINGLE_PLUGIN == plugin_name
-
-        return not self.config.active_plugins or plugin_name in self.config.active_plugins
-
     def start_subscribing(self):
-        log.debug('Starting message subscription')
+        log.info('Starting message subscription')
 
         # Setup sqAPI general exchange listener
         log.debug('Starting Exchange Listener')
@@ -103,7 +59,7 @@ class ProcessManager:
             process_pool = [
                 multiprocessing.Process(target=plugin.execute, args=[
                     plugin.config, plugin.database, message.body, metadata, open(data_path, 'rb')
-                ]) for plugin in self.plugins
+                ]) for plugin in self.plugin_manager.plugins
                 if valid_data_type(message, plugin)
             ]
 
@@ -134,9 +90,12 @@ class ProcessManager:
         log.debug('Validating mime type')
         msg_type = message.get('data_type', 'UNKNOWN')
         log.debug('Message Mime type: {}'.format(msg_type))
-        log.debug('Accepted Mime types: {}'.format(self.accepted_types))
+        log.debug('Accepted Mime types: {}'.format(self.plugin_manager.accepted_types))
 
-        if self.accepted_types and (msg_type not in self.accepted_types and '*' not in self.accepted_types):
+        if self.plugin_manager.accepted_types and (
+                msg_type not in self.plugin_manager.accepted_types
+                and '*' not in self.plugin_manager.accepted_types
+        ):
             err = 'Mime type {} is not supported by any of the active sqAPI plugins'.format(msg_type)
             log.debug(err)
             raise NotImplementedError(err)
